parts 4.20-4.22

+++ part4/part4blogs/controllers/blogs.js
@@ -4,13 +4,6 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-      return authorization.substring(7)
-    }
-    return null
-  }
 
 blogRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id:1})
@@ -19,13 +12,10 @@ blogRouter.get('/', async (request, response) => {
 
 blogRouter.post('/', async (request, response, next) => {
     const body = request.body
-    const token = getTokenFrom(request)
-    const decodedToken = jwt.verify(token, process.env.SECRET)
-    if (!decodedToken.id) {
-        return response.status(401).json({ error: 'token missing or invalid' })
-    }
-    const users = await User.find({})
-    const currentUser = users.filter(u=>u.username===body.username)[0]
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    const user = await User.findById(decodedToken.id)
     
     if(!body.title||!body.url)response.status(400).json({error:'provide a title and url'})
 
@@ -40,15 +30,15 @@ blogRouter.post('/', async (request, response, next) => {
         author: body.author,
         url: body.url,
         likes: body.likes,
-        user: currentUser._id
+        user: user._id
     })
 
     try {
         const savedBlog = await blog.save()
         logger.info(`added ${blog.title} to the blog list`)
-        currentUser.blogs = currentUser.blogs.concat(savedBlog._id)
-        await currentUser.save()
-        logger.info(`blog linked to user ${currentUser.username}`)
+        user.blogs = user.blogs.concat(savedBlog._id)
+        await user.save()
+        logger.info(`blog linked to user ${user.username}`)
         response.json(savedBlog.toJSON())
     } catch(exception) {
         next(exception)
@@ -56,12 +46,24 @@ blogRouter.post('/', async (request, response, next) => {
 })
 
 blogRouter.delete('/:id', async (request, response, next) => {
+
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    const user = await User.findById(decodedToken.id)
+    const blogToDelete = await Blog.findById(request.params.id)
+    //check if user objects are equal. This prevents other users from deleting your content
+    if ( blogToDelete.user._id.toString() === user._id.toString() ) {
     try {
       await Blog.findByIdAndRemove(request.params.id)
       response.status(204).end()
-    } catch (exception) {
+    } 
+    catch (exception) {
       next(exception)
     }
+    } else {
+        return response.status(401).json({ error: `Unauthorized` })
+    }
+
   })
 
 blogRouter.put('/:id', async (request, response, next) => {
@@ -71,6 +73,13 @@ blogRouter.put('/:id', async (request, response, next) => {
         body.likes = 0
     }
 
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    const user = await User.findById(decodedToken.id)
+    const blogToUpdate = await Blog.findById(request.params.id)
+
+    
+    if ( blogToUpdate.user._id.toString() === user._id.toString() ) {
     const blog = {
         title: body.title,
         author: body.author,
@@ -84,6 +93,9 @@ blogRouter.put('/:id', async (request, response, next) => {
         response.json(updatedBlog.toJSON())
     } catch (exception) {
         next(exception)
+    }
+    } else {
+        return response.status(401).json({ error: `Unauthorized` })
     }
 })
